Conflicts:
	DB/pom.xml

+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/KogakuGassanShinseishoKanyurekiIdentifier.java
@@ -24,7 +24,7 @@ public class KogakuGassanShinseishoKanyurekiIdentifier implements Serializable {
     private final HokenshaNo 保険者番号;
     private final RString 整理番号;
     private final RString 加入歴番号;
-    private final Decimal 履歴番号;
+    private final int 履歴番号;
 
     /**
      * コンストラクタです。
@@ -41,7 +41,7 @@ public class KogakuGassanShinseishoKanyurekiIdentifier implements Serializable {
             HokenshaNo 保険者番号,
             RString 整理番号,
             RString 加入歴番号,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         this.被保険者番号 = 被保険者番号;
         this.対象年度 = 対象年度;
         this.保険者番号 = 保険者番号;
