+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/KogakuJuryoininKeiyakuJigyoshaBuilder.java
@@ -70,7 +70,7 @@ public class KogakuJuryoininKeiyakuJigyoshaBuilder {
      * @param 履歴番号 履歴番号
      * @return {@link KogakuJuryoininKeiyakuJigyoshaBuilder}
      */
-    public KogakuJuryoininKeiyakuJigyoshaBuilder set履歴番号(Decimal 履歴番号) {
+    public KogakuJuryoininKeiyakuJigyoshaBuilder set履歴番号(int 履歴番号) {
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
         entity.setRirekiNo(履歴番号);
         return this;
