+++ DB/DBC/standard/business/src/test/java/jp/co/ndensan/reams/db/dbc/business/core/basic/RiyoshaFutanWariaiMeisaiIdentifierTest.java
@@ -27,8 +27,8 @@ public class RiyoshaFutanWariaiMeisaiIdentifierTest extends DbcTestBase {
 //TODO 主キーの数が足りない場合、追加してください。
     private static FlexibleYear 年度;
     private static HihokenshaNo 被保険者番号;
-    private static Decimal 履歴番号;
-    private static Decimal 枝番号;
+    private static int 履歴番号;
+    private static int 枝番号;
 
     @BeforeClass
     public static void setUpClass() {
